♻️move eventRateLimiter to domain
since it manipulates RawError, it should be on domain
