+++ packages/core/src/index.ts
@@ -61,7 +61,7 @@ export * from './tools/urlPolyfill'
 export * from './tools/timeUtils'
 export * from './tools/utils'
 export * from './tools/sanitize'
-export * from './tools/createEventRateLimiter'
+export * from './domain/eventRateLimiter/createEventRateLimiter'
 export * from './tools/browserDetection'
 export { sendToExtension } from './tools/sendToExtension'
 export { runOnReadyState } from './tools/runOnReadyState'
